backend/requirements.txt
fastapi
uvicorn
python-docx
pandas
